backend/requirements.txt
alembic==1.14.1
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.6
django-admin-interface==0.29.4
django-colorfield==0.12.0
djangorestframework==3.15.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.7.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==0.20.0
python-slugify==8.0.4
six==1.17.0
SQLAlchemy==1.4.36
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
Werkzeug==2.3.7
